Don't skip update logic on Streakable record creation
This handles the case where a new Streakable record is created with only partial
attributes (ex. you create a new Club with a name & address, but latitude and
longitude aren't set -- a callback will geocode the address to set the latitude
and longitude and we should send those updates to Streak).